# Refactor make_downshift_arrays and enhance the performance of topology construction