Fixes #3721 

Changes made in this Pull Request:
 - refactor make_downshift_arrays
 - lazy building `Universe._topology.tt. _RA, _SR` during construction and serialization.


PR Checklist
------------
 - [x] Tests?
 - [x] Docs?
 - [x] CHANGELOG updated?
 - [x] Issue raised/referenced?