@pep8speaks 2022-06-19
Hello @yuxuanzhuang! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`package/MDAnalysis/core/topology.py`](https://github.com/MDAnalysis/mdanalysis/blob/71920b1cd9a60171bb2790fcc1ec5ed9abff17d3/package/MDAnalysis/core/topology.py):

> [Line 427:29](https://github.com/MDAnalysis/mdanalysis/blob/71920b1cd9a60171bb2790fcc1ec5ed9abff17d3/package/MDAnalysis/core/topology.py#L427): [E128](https://duckduckgo.com/?q=pep8%20E128) continuation line under-indented for visual indent