@yuxuanzhuang 2022-06-23
**Performance Improvement**
```python
from MDAnalysis.tests.datafiles import PDB
import MDAnalysis as mda
from MDAnalysis.core.topology import (
    TransTable,
    make_downshift_arrays,
)

u = mda.Universe(PDB)
n_atoms = u._topology.tt.n_atoms
# 47681 atoms
n_residues = u._topology.tt.n_residues
# 11302 residues
n_segments = u._topology.tt.n_segments
# 1 segment
AR = u._topology.tt._AR
RS = u._topology.tt._RS

%timeit make_downshift_arrays(AR, n_residues)
# original
29.6 ms ± 254 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# now
12.1 ms ± 73.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

%timeit TransTable(n_atoms, n_residues, n_segments, AR, RS)
# original 
29.8 ms ± 329 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
# now
17 µs ± 57.4 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

tt = TransTable(n_atoms, n_residues, n_segments, AR, RS)
%timeit pickle.loads(pickle.dumps(tt))
# original
29.6 ms ± 1.27 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
# now
69.7 µs ± 740 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```